Reverse PathPlanner paths
The robot will always start facing the grid, so that we can score any
preloads. However, the PathPlanner (correctly) assumes that the front of
the robot is where the arm is. Thus, we need to reverse all the paths so
that instead of rotating 180° at the start of the path (which could cause
collisions or other unexpected behavior), the robot just drives the path
facing backwards.